2 - FIX int for unsigned int

+++ 0x10-variadic_functions/2-print_strings.c
@@ -7,19 +7,19 @@
  */
 void print_strings(const char *separator, const unsigned int n, ...)
 {
-	int i;
+	unsigned int i;
 	va_list ap;
 	char *stri;
 
 	va_start(ap, n);
-	for (i = 0; i < (int)n; i++)
+	for (i = 0; i < n; i++)
 	{
 		stri = va_arg(ap, char*);
 		if (separator == NULL)
 			printf("(nil)");
 		else
 			printf("%s", stri);
-		if (separator != NULL && i < ((int)n - 1))
+		if (separator != NULL && i < (n - 1))
 			printf("%s", separator);
 	}
 	printf("\n");
